Move instance checkpoint capture and restore into atomic_lib
