+++ CHANGELOG.md
@@ -7,6 +7,9 @@ See [STATUS.md](server/STATUS.md) to learn more about which features will remain
 
 ## UNRELEASED
 
+- Instance checkpoint capture, verification and offline restore are available in
+  `atomic_lib` through the optional native `backup` feature, without server/Actix.
+
 - Add opt-in full-instance backups: `--backup-dir`, local authenticated backup
   control, `backup`/`restore` CLI commands, checksummed ZIP64 archives and an
   offline restore guard. Writes and incoming sync application pause during
