+++ lib/Cargo.toml
@@ -72,6 +72,11 @@ wasm-bindgen = { version = "0.2.122", optional = true }
 wasm-bindgen-futures = { version = "0.4.72", optional = true }
 web-sys = { version = "0.3.99", optional = true, features = ["DomException", "FileSystemDirectoryHandle", "FileSystemFileHandle", "FileSystemGetFileOptions", "FileSystemSyncAccessHandle", "FileSystemReadWriteOptions", "StorageManager", "WorkerGlobalScope", "WorkerNavigator"] }
 
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+chrono = { version = "0.4.44", optional = true }
+tempfile = { version = "3", optional = true }
+zip = { version = "8.6.0", default-features = false, features = ["deflate"], optional = true }
+
 [dev-dependencies]
 criterion = { version = "0.8.2", features = ["async_tokio"] }
 iai = "0.1"
@@ -86,6 +91,8 @@ optional = true
 version = "0.32"
 
 [features]
+## Native full-instance checkpoints; no hosted server or Actix dependency.
+backup = ["db-redb", "dep:chrono", "dep:tempfile", "dep:zip", "tokio/time"]
 config = ["directories", "toml"]
 ## Core Db with encoding support (BTreeMapStore included).
 ## Does NOT pull in sled — works in WASM.
