+++ server/Cargo.toml
@@ -84,7 +84,6 @@ kuchikiki = "0.8.2"
 indexmap = { version = "2.14.0", features = ["std"] }
 lol_html = "2.9"
 zip = { version = "8.6.0" }
-tempfile = "3"
 reqwest = { version = "0.13.3", default-features = false, features = [
   "rustls",
   "json",
@@ -167,7 +166,7 @@ version = "4.13.0"
 # `db-sled` is included so an in-place upgrade from a pre-redb (sled) store
 # auto-migrates to redb on first boot. It is inert unless a legacy sled DB is
 # present, so it adds no runtime cost for fresh installs.
-features = ["config", "db-redb", "db-sled", "rdf", "discovery", "iroh", "ring", "ws"]
+features = ["backup", "config", "db-redb", "db-sled", "rdf", "discovery", "iroh", "ring", "ws"]
 path = "../lib"
 version = "0.41.0-beta.6"
 
@@ -201,7 +200,7 @@ features = ["env-filter"]
 version = "0.3.23"
 
 [dev-dependencies]
-redb = "4.1.0"
+tempfile = "3"
 actix-rt = "2.11.0"
 assert_cmd = "2.2.2"
 mainline = "6.2.0"
